% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNA4tutti-package.r
\docType{package}
\name{sna4tutti}
\alias{sna4tutti-package}
\alias{sna4tutti}
\title{SNA4DS}
\description{
Package supporting the SNA4DS course
}
\details{
Functions, datasets, and tutorials, useful for the SNA4DS course.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SNAnalyst/sna4tutti}
  \item Report bugs at \url{https://github.com/SNAnalyst/sna4tutti/issues}
}

}
\author{
\strong{Maintainer}: Roger Leenders \email{SNA4DS@jads.nl}

Authors:
\itemize{
  \item Claudia Zucca \email{SNA4DS@jads.nl}
}

}
\keyword{package}
