% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broken_checks.R
\name{broken_info}
\alias{broken_info}
\title{Within tutorial sna4tutti unit test error report function}
\usage{
broken_info(broken_list = list())
}
\value{
nothing relevant, this function is useful for its side effect.
}
\description{
This function prints an error message in case of broken code chunk and it
reports the name of the broken code box
}
\examples{
\dontrun{
broken_info()
}
}
