% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNA4tutti-package.r
\docType{package}
\name{sna4tutti}
\alias{sna4tutti}
\title{SNA4DS}
\description{
Package supporting the SNA4DS course
}
\details{
Functions, datasets, and tutorials, useful for the SNA4DS course.
}
\keyword{package}
